Update aseqjoy.c - add button mapping
Button Mapping
--------------

You can now map specific buttons to MIDI controllers using the `-b` option. Only the buttons specified with the `-b` option will be included in the button map. Buttons not defined in the command will be ignored. The format for the `-b` option is `button=controller`.

Example usage:

```sh
./aseqjoy -d 0 -v -0 1 -1 2 -2 3 -3 4 -b 0=20 -b 1=21 -b 2=22
```

Explanation:

-d 0: Select joystick 0.
-v: Enable verbose mode.
-0 1, -1 2, -2 3, -3 4: Map joystick axes to MIDI controllers 1, 2, 3, and 4 respectively.
-b 0=20: Map button 0 to MIDI controller 20.
-b 1=21: Map button 1 to MIDI controller 21.
-b 2=22: Map button 2 to MIDI controller 22.

With this configuration, only buttons 0, 1, and 2 will be mapped to MIDI controllers 20, 21, and 22, respectively. Other buttons will be ignored.